Separate string-based dv dicts into files, for easier exclusion of unused code
